backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Cors==5.0.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
MarkupSafe==3.0.2
numpy==2.1.3
openai==1.55.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pluggy==1.5.0
psycopg2-binary
pydantic==2.10.2
pydantic_core==2.27.1
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
stripe==11.3.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
dotenv
pathlib

